+++ source/extensions/lightspeed.trex.stage_manager.plugin.interaction.usd/lightspeed/trex/stage_manager/plugin/interaction/usd/all_meshes.py
@@ -15,6 +15,7 @@
 * limitations under the License.
 """
 
+from lightspeed.trex.utils.common.prim_utils import get_extended_selection as _get_extended_selection
 from omni.flux.stage_manager.factory.plugins import StageManagerFilterPlugin as _StageManagerFilterPlugin
 from omni.flux.stage_manager.factory.plugins import StageManagerTreePlugin as _StageManagerTreePlugin
 from omni.flux.stage_manager.plugin.interaction.usd.base import (
@@ -47,3 +48,6 @@ class AllMeshesInteractionPlugin(_StageManagerUSDInteractionPlugin):
         "IsVisibleActionWidgetPlugin",
         "PrimTreeWidgetPlugin",
     ]
+
+    def _get_selection(self):
+        return _get_extended_selection(self._context_name)