+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Created 1.0.0 build
 - Added release notes for 1.0.0
 - Added tests for REST API endpoints
+- REMIX-3891: Added Material Property Widget control with Stage Manager Materials Tab
 
 ### Changed
 - Improved the documentation layout and contents