+++ source/extensions/lightspeed.trex.properties_pane.shared.asset_replacements.widget/docs/CHANGELOG.md
@@ -1,6 +1,12 @@
 # Changelog
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
+## [1.4.0]
+## Changed
+- Material property widget & pinning now driven by USD stage selection
+- Direct material prim selection now populates the material property widget
+- Fixed test bug where tests couldn't re-run without failing
+
 ## [1.3.2]
 ## Changed
 - Update to Kit 106.5