+++ source/extensions/lightspeed.trex.selection_tree.shared.widget/docs/CHANGELOG.md
@@ -1,6 +1,12 @@
 # Changelog
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
+## [1.5.3]
+## Changed
+- Changed `__get_model_from_prototype_path()` to exclude prims that are simultaneously materials
+- The Selection Tree will no longer activate solely with mesh prims that are also materials
+- Moved `__reference_file_paths()` to generic `prim_utils.py` from `model.py`
+
 ## [1.5.2]
 ## Changed
 - Update to Kit 106.5