+++ source/extensions/lightspeed.trex.stage_manager.plugin.interaction.usd/docs/CHANGELOG.md
@@ -1,6 +1,11 @@
 # Changelog
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
+## [1.2.0]
+### Added
+- Added `AllMaterialsRemixInteractionPlugin` to wrap flux side
+- Added `_get_selection()` overrides for category and mesh tabs to fix scroll centering
+
 ## [1.1.1]
 ### Fixed
 - Added missing compatible widgets